fix: break apps/web barrel cycles (M1/M2) + de-hardcode branch-reply pluginId (A4)
- M1: extract AssetRender into its own module so asset-render/index is a pure
  barrel (AssetTurnSidebar no longer imports the barrel back).
- M2: app-kv-store imports StatePatchRecord from api/types instead of the
  api.ts barrel, breaking the api -> overlay -> app-kv-store cycle.
  madge: 0 circular deps in apps/web.
- A4 (part 1): BranchReplyBlock resolves authoring pluginId from message /
  block data instead of injecting the literal "branch-reply".

A4 block-type dispatch migration (branch-reply -> ui-spec) and chat-list
virtualization deferred — both need running-browser verification.
